requirements.txt
# Pinned runtime deps (in sync with pyproject.toml)
dnspython==2.7.0
httpx[http2,socks]==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
python-telegram-bot[job-queue]==22.3
PyYAML==6.0.2
tldextract==5.3.0
icmplib==3.0 ; platform_system != "Windows"